manda_project/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
coverage==7.3.2
cssutils==2.9.0
Django==4.2.6
djangorestframework==3.14.0
drf-yasg==1.21.7
future==0.18.3
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
protobuf==4.24.4
pynliner==0.8.0
pytz==2023.3.post1
PyYAML==6.0.1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
